Đăng ký, Đăng nhập, Phân quyền, Quản lý thông tin, Xuất hóa đơn

+++ QLBanDoAnNhanh/QLBanDoAnNhanh/Models/ChiTietDonHang.cs
@@ -13,7 +13,7 @@ public partial class ChiTietDonHang
 
     public int SoLuong { get; set; }
 
-    public int TongTien { get; set; }
+    public double TongTien { get; set; }
 
     public virtual DonHang MaDhNavigation { get; set; }
